Refactor Overcooked V3 step pipeline (#23)
* Refactor Overcooked V3 step pipeline

* Clean step-system file endings
