+++ jaxmarl/environments/overcooked_v3/overcooked.py
@@ -1,192 +1,75 @@
-"""Overcooked V3 Environment with pot burning, order queue, and conveyor belts."""
+"""Public Overcooked V3 environment wrapper."""
 
-from enum import Enum
-from typing import List, Optional, Union, Tuple, Dict
+from typing import Dict, List, Optional, Tuple, Union
 import warnings
-import numpy as np
+
+import chex
 import jax
 import jax.numpy as jnp
-from jax import lax
-import chex
+import numpy as np
 
 from jaxmarl.environments import MultiAgentEnv
 from jaxmarl.environments import spaces
-from jaxmarl.environments.overcooked_v3.common import (
-    ACTION_TO_DIRECTION,
-    DIR_TO_VEC,
-    MAX_INGREDIENTS,
-    Actions,
-    StaticObject,
-    DynamicObject,
-    Direction,
-    ButtonAction,
-    Position,
-    Agent,
+from jaxmarl.environments.overcooked_v3.agent_step import (
+    barriers_occupied,
+    is_agent_walkable,
+    run_agent_action_phase,
 )
-from jaxmarl.environments.overcooked_v3.layouts import overcooked_v3_layouts, Layout
+from jaxmarl.environments.overcooked_v3.common import (
+    Actions,
+    Agent,
+    ButtonAction,
+    StaticObject,
+)
+from jaxmarl.environments.overcooked_v3.config import OvercookedV3Config
+from jaxmarl.environments.overcooked_v3.initialization import (
+    randomize_agent_positions,
+    randomize_state,
+    sample_recipe,
+)
+from jaxmarl.environments.overcooked_v3.interactions import process_interact
+from jaxmarl.environments.overcooked_v3.layouts import Layout, overcooked_v3_layouts
+from jaxmarl.environments.overcooked_v3.observations import (
+    calculate_observation_shape,
+    get_obs,
+    get_obs_default,
+    get_obs_for_type,
+)
+from jaxmarl.environments.overcooked_v3.reset import reset_overcooked_v3
 from jaxmarl.environments.overcooked_v3.settings import (
-    DELIVERY_REWARD,
-    POT_COOK_TIME,
-    POT_BURN_TIME,
-    ORDER_EXPIRED_PENALTY,
-    DEFAULT_ORDER_GENERATION_RATE,
-    DEFAULT_ORDER_EXPIRATION_TIME,
+    DEFAULT_BARRIER_DURATION,
     DEFAULT_MAX_ORDERS,
-    SHAPED_REWARDS,
-    MAX_POTS,
-    MAX_ITEM_CONVEYORS,
-    MAX_PLAYER_CONVEYORS,
-    MAX_MOVING_WALLS,
+    DEFAULT_ORDER_EXPIRATION_TIME,
+    DEFAULT_ORDER_GENERATION_RATE,
+    DELIVERY_REWARD,
+    MAX_BARRIERS,
     MAX_BUTTONS,
     MAX_BUTTON_TARGETS,
-    MAX_BARRIERS,
+    MAX_ITEM_CONVEYORS,
+    MAX_MOVING_WALLS,
+    MAX_PLAYER_CONVEYORS,
+    MAX_POTS,
     MAX_PRESSURE_PLATES,
-    DEFAULT_BARRIER_DURATION,
+    POT_BURN_TIME,
+    POT_COOK_TIME,
 )
-from jaxmarl.environments.overcooked_v3.utils import (
-    tree_select,
-    compute_enclosed_spaces,
+from jaxmarl.environments.overcooked_v3.state import ObservationType, State
+from jaxmarl.environments.overcooked_v3.step import is_terminal, step_overcooked_v3
+from jaxmarl.environments.overcooked_v3.systems.barriers import (
+    update_barrier_timers,
+    update_pressure_plates,
 )
-
-
-# =============================================================================
-# Observation Types and State Container
-# =============================================================================
-
-
-class ObservationType(str, Enum):
-    DEFAULT = "default"
-    FEATURIZED = "featurized"
-
-
-@chex.dataclass
-class State:
-    """Environment state for Overcooked V3."""
-
-    agents: Agent
-
-    # Grid: height x width x 3 channels
-    # Channel 0: static objects
-    # Channel 1: dynamic items (plates, ingredients, soups)
-    # Channel 2: extra info (pot timers, conveyor directions)
-    grid: chex.Array
-
-    # Pot state (fixed size arrays for JIT compatibility)
-    # pot_positions stores (y, x) for each pot, pot_active_mask indicates valid pots
-    pot_positions: chex.Array  # [max_pots, 2] - (y, x) positions
-    pot_cooking_timer: (
-        chex.Array
-    )  # [max_pots] - countdown to cooked (0 when idle/cooked)
-    pot_active_mask: chex.Array  # [max_pots] - bool, which pot slots are valid
-
-    # Order queue state (optional feature)
-    order_types: chex.Array  # [max_orders] - SoupType enum values
-    order_expirations: chex.Array  # [max_orders] - steps remaining
-    order_active_mask: chex.Array  # [max_orders] - bool, which order slots are valid
-
-    # Item conveyor state
-    item_conveyor_positions: chex.Array  # [max_item_conveyors, 2] - (y, x)
-    item_conveyor_directions: chex.Array  # [max_item_conveyors] - Direction enum
-    item_conveyor_active_mask: chex.Array  # [max_item_conveyors] - bool
-
-    # Player conveyor state
-    player_conveyor_positions: chex.Array  # [max_player_conveyors, 2] - (y, x)
-    player_conveyor_directions: chex.Array  # [max_player_conveyors] - Direction enum
-    player_conveyor_active_mask: chex.Array  # [max_player_conveyors] - bool
-
-    # Moving wall state
-    moving_wall_positions: chex.Array  # [max_moving_walls, 2] - (y, x)
-    moving_wall_directions: chex.Array  # [max_moving_walls] - Direction enum
-    moving_wall_active_mask: chex.Array  # [max_moving_walls] - bool
-    moving_wall_paused: chex.Array  # [max_moving_walls] - bool
-    moving_wall_bounce: chex.Array  # [max_moving_walls] - bool
-
-    # Button state
-    button_positions: chex.Array  # [max_buttons, 2] - (y, x)
-    button_target_idxs: chex.Array  # [max_buttons, max_button_targets]
-    button_target_mask: chex.Array  # [max_buttons, max_button_targets] - bool
-    button_action_type: chex.Array  # [max_buttons] - ButtonAction enum
-    button_active_mask: chex.Array  # [max_buttons] - bool
-    button_toggled: chex.Array  # [max_buttons] - bool (current toggle state)
-
-    # Barrier state
-    barrier_positions: chex.Array  # [max_barriers, 2] - (y, x)
-    barrier_active: chex.Array  # [max_barriers] - bool (blocks when True)
-    barrier_active_mask: chex.Array  # [max_barriers] - bool (which slots are valid)
-    barrier_timer: (
-        chex.Array
-    )  # [max_barriers] - steps until reactivation (0 means permanent state)
-    barrier_duration: (
-        chex.Array
-    )  # [max_barriers] - configured duration for timed deactivation
-
-    # Pressure Plate State
-    pressure_plate_positions: chex.Array  # [max_pressure_plates, 2] - (y, x)
-    pressure_plate_linked_barrier: chex.Array  # [max_pressure_plates, max_barriers] - linked barrier mask
-    pressure_plate_action_type: chex.Array  # [max_pressure_plates] - ButtonAction enum
-    pressure_plate_active_mask: chex.Array  # [max_pressure_plates] - bool (which slots are valid)
-    pressure_plate_toggled: chex.Array  # [max_pressure_plates] - bool (currently pressed)
-
-    # Episode state
-    time: chex.Array
-    terminal: bool
-    recipe: int  # Current target recipe (bit-encoded)
-
-    # Delivery tracking
-    new_correct_delivery: bool = False
-
-
-# =============================================================================
-# Overcooked V3 Environment
-# =============================================================================
-
+from jaxmarl.environments.overcooked_v3.systems.conveyors import (
+    move_items_on_item_conveyors,
+    push_players_on_player_conveyors,
+)
+from jaxmarl.environments.overcooked_v3.systems.moving_walls import move_moving_walls
+from jaxmarl.environments.overcooked_v3.systems.orders import process_order_queue
+from jaxmarl.environments.overcooked_v3.systems.pots import update_pot_timers
+from jaxmarl.environments.overcooked_v3.utils import compute_enclosed_spaces
 
 class OvercookedV3(MultiAgentEnv):
-    """Overcooked V3 environment with pot burning, order queue, and conveyors.
-
-    Methods:
-        reset(key) -> Tuple[Dict[str, Array], State]:
-            Reset the environment and return initial observations and state.
-
-        step_env(key, state, actions) -> Tuple[obs, State, rewards, dones, info]:
-            Perform a single timestep: process actions, conveyors, orders, and check termination.
-
-        step_agents(key, state, actions) -> Tuple[State, float, Array]:
-            Process agent movement (with collision resolution) and interact actions.
-
-        process_interact(grid, agent, all_inventories, recipe, pot_timers,
-            pot_positions, pot_active_mask) -> Tuple[grid, agent, correct_delivery,
-            reward, shaped_reward, pot_timers]:
-            Handle a single agent's interact action (pickup, drop, cook, deliver).
-
-        is_terminal(state) -> bool:
-            Check whether the episode is done (max steps reached).
-
-        get_obs(state) -> Dict[str, Array]:
-            Get observations for all agents, dispatching by per-agent observation type.
-
-        get_obs_for_type(state, obs_type) -> Dict[str, Array]:
-            Get observations for a specific observation type (default or featurized).
-
-        get_obs_default(state) -> Array:
-            Build default grid-based observation tensors for all agents.
-
-        name (property) -> str:
-            Return the environment name.
-
-        num_actions (property) -> int:
-            Return the number of possible actions.
-
-        action_space(agent_id) -> spaces.Discrete:
-            Return the discrete action space.
-
-        observation_space(agent_id) -> spaces.Box:
-            Return the box observation space.
-    """
-
-    # -------------------------------------------------------------------------
-    # Construction and Layout Preprocessing
-    # -------------------------------------------------------------------------
+    """Overcooked V3 environment backed by explicit functional JAX logic."""
 
     def __init__(
         self,
@@ -252,7 +135,8 @@ class OvercookedV3(MultiAgentEnv):
         if isinstance(layout, str):
             if layout not in overcooked_v3_layouts:
                 raise ValueError(
-                    f"Invalid layout: {layout}, allowed layouts: {list(overcooked_v3_layouts.keys())}"
+                    f"Invalid layout: {layout}, "
+                    f"allowed layouts: {list(overcooked_v3_layouts.keys())}"
                 )
             layout = overcooked_v3_layouts[layout]
         elif not isinstance(layout, Layout):
@@ -382,7 +266,13 @@ class OvercookedV3(MultiAgentEnv):
         )
 
         # Compute observation shape
-        self.obs_shape = self._get_obs_shape()
+        self.obs_shape = calculate_observation_shape(
+            self.width,
+            self.height,
+            self.layout,
+            self.observation_type,
+            self.agent_view_size,
+        )
 
         # Extract pot positions from layout
         pot_mask = layout.static_objects == StaticObject.POT
@@ -469,7 +359,6 @@ class OvercookedV3(MultiAgentEnv):
                         target_idx = self._button_target_idxs[button_idx, target_slot]
                         self._moving_wall_initial_paused[target_idx] = True
 
-
         # Extract barrier info from layout
         self._barrier_positions = np.zeros((MAX_BARRIERS, 2), dtype=np.int32)
         self._barrier_initial_active = np.zeros(MAX_BARRIERS, dtype=bool)
@@ -477,7 +366,9 @@ class OvercookedV3(MultiAgentEnv):
         self._barrier_duration_config = np.zeros(MAX_BARRIERS, dtype=np.int32)
 
         # Extract pressure plate info from layout
-        self._pressure_plate_positions = np.zeros((MAX_PRESSURE_PLATES, 2), dtype=np.int32)
+        self._pressure_plate_positions = np.zeros(
+            (MAX_PRESSURE_PLATES, 2), dtype=np.int32
+        )
         self._pressure_plate_linked_barrier = np.zeros(
             (MAX_PRESSURE_PLATES, MAX_BARRIERS), dtype=bool
         )
@@ -494,7 +385,6 @@ class OvercookedV3(MultiAgentEnv):
             self._pressure_plate_action_type[i] = action_type
             self._pressure_plate_active_mask[i] = True
 
-        num_barriers = len(layout.barrier_info)
         for i, (y, x, active) in enumerate(layout.barrier_info[:MAX_BARRIERS]):
             self._barrier_positions[i] = [y, x]
             self._barrier_initial_active[i] = active
@@ -509,234 +399,73 @@ class OvercookedV3(MultiAgentEnv):
             else:
                 self._barrier_duration_config[i] = barrier_duration
 
-    def _get_obs_shape(self) -> Tuple[int, ...]:
-        """Calculate observation shape based on observation type."""
-        if self.agent_view_size:
-            view_size = self.agent_view_size * 2 + 1
-            view_width = min(self.width, view_size)
-            view_height = min(self.height, view_size)
-        else:
-            view_width = self.width
-            view_height = self.height
+        self.config = self._build_config()
+        self._step_jit = jax.jit(step_overcooked_v3, static_argnames=("config",))
 
-        def _get_obs_shape_single(obs_type):
-            if obs_type == ObservationType.DEFAULT:
-                num_ingredients = self.layout.num_ingredients
-                # Layers breakdown:
-                # - agent_layer: 1 (pos) + 4 (dir) + (2 + num_ing) (inv) = 7 + num_ing
-                # - other_agent_layers: same = 7 + num_ing
-                # - static_layers: 11 (wall, goal, pot, recipe, plate, item_conv, player_conv, moving_wall, button, barrier, pressure_plate)
-                # - ingredient_pile_layers: num_ing
-                # - ingredients_layers: 2 + num_ing
-                # - recipe_layers: 2 + num_ing
-                # - extra_layers: 1 (pot timer)
-                # Total: 30 + 5 * num_ingredients
-                num_layers = 30 + 5 * num_ingredients
-                return (view_height, view_width, num_layers)
-            elif obs_type == ObservationType.FEATURIZED:
-                # Simplified feature vector
-                return (64,)  # Placeholder
-            else:
-                raise ValueError(f"Invalid observation type: {obs_type}")
-
-        if isinstance(self.observation_type, list):
-            return [
-                _get_obs_shape_single(obs_type) for obs_type in self.observation_type
-            ]
-
-        return _get_obs_shape_single(self.observation_type)
-
-    # -------------------------------------------------------------------------
-    # Reset and Random Initialization
-    # -------------------------------------------------------------------------
+    def _build_config(self) -> OvercookedV3Config:
+        """Build the static configuration object passed through functional logic."""
+        return OvercookedV3Config(
+            height=self.height,
+            width=self.width,
+            layout=self.layout,
+            num_agents=self.num_agents,
+            agents=self.agents,
+            action_set=self.action_set,
+            observation_type=self.observation_type,
+            agent_view_size=self.agent_view_size,
+            obs_shape=self.obs_shape,
+            max_steps=self.max_steps,
+            pot_cook_time=self.pot_cook_time,
+            pot_burn_time=self.pot_burn_time,
+            enable_order_queue=self.enable_order_queue,
+            max_orders=self.max_orders,
+            order_generation_rate=self.order_generation_rate,
+            order_expiration_time=self.order_expiration_time,
+            enable_item_conveyors=self.enable_item_conveyors,
+            enable_player_conveyors=self.enable_player_conveyors,
+            enable_moving_walls=self.enable_moving_walls,
+            enable_buttons=self.enable_buttons,
+            enable_pressure_plates=self.enable_pressure_plates,
+            delivery_reward=self.delivery_reward,
+            shaped_rewards_enabled=self.shaped_rewards_enabled,
+            random_reset=self.random_reset,
+            random_agent_positions=self.random_agent_positions,
+            possible_recipes=self.possible_recipes,
+            enclosed_spaces=self.enclosed_spaces,
+            pot_positions=self._pot_positions,
+            pot_active_mask=self._pot_active_mask,
+            item_conveyor_positions=self._item_conveyor_positions,
+            item_conveyor_directions=self._item_conveyor_directions,
+            item_conveyor_active_mask=self._item_conveyor_active_mask,
+            player_conveyor_positions=self._player_conveyor_positions,
+            player_conveyor_directions=self._player_conveyor_directions,
+            player_conveyor_active_mask=self._player_conveyor_active_mask,
+            moving_wall_positions=self._moving_wall_positions,
+            moving_wall_directions=self._moving_wall_directions,
+            moving_wall_active_mask=self._moving_wall_active_mask,
+            moving_wall_initial_paused=self._moving_wall_initial_paused,
+            moving_wall_bounce=self._moving_wall_bounce,
+            button_positions=self._button_positions,
+            button_target_idxs=self._button_target_idxs,
+            button_target_mask=self._button_target_mask,
+            button_action_type=self._button_action_type,
+            button_active_mask=self._button_active_mask,
+            barrier_positions=self._barrier_positions,
+            barrier_initial_active=self._barrier_initial_active,
+            barrier_active_mask=self._barrier_active_mask,
+            barrier_duration_config=self._barrier_duration_config,
+            pressure_plate_positions=self._pressure_plate_positions,
+            pressure_plate_linked_barrier=self._pressure_plate_linked_barrier,
+            pressure_plate_action_type=self._pressure_plate_action_type,
+            pressure_plate_active_mask=self._pressure_plate_active_mask,
+        )
 
     def reset(
         self,
         key: chex.PRNGKey,
     ) -> Tuple[Dict[str, chex.Array], State]:
-        """Reset the environment."""
-        layout = self.layout
-
-        # Initialize grid
-        static_objects = layout.static_objects
-        grid = jnp.stack(
-            [
-                static_objects,
-                jnp.zeros_like(static_objects),  # dynamic items
-                jnp.zeros_like(
-                    static_objects
-                ),  # extra info (pot timers, conveyor dirs)
-            ],
-            axis=-1,
-            dtype=jnp.int32,
-        )
-
-        # Store conveyor directions in extra channel
-        for i, (y, x, direction) in enumerate(layout.item_conveyor_info):
-            grid = grid.at[y, x, 2].set(direction)
-        for i, (y, x, direction) in enumerate(layout.player_conveyor_info):
-            grid = grid.at[y, x, 2].set(direction)
-
-        # Store moving wall directions in extra channel
-        for y, x, direction, _bounce in layout.moving_wall_info:
-            grid = grid.at[y, x, 2].set(direction)
-
-        # Initialize agents
-        num_agents = self.num_agents
-        x_positions, y_positions = map(jnp.array, zip(*layout.agent_positions))
-        agents = Agent(
-            pos=Position(x=x_positions, y=y_positions),
-            dir=jnp.full((num_agents,), Direction.UP),
-            inventory=jnp.zeros((num_agents,), dtype=jnp.int32),
-        )
-
-        # Sample recipe
-        key, subkey = jax.random.split(key)
-        recipe = self._sample_recipe(subkey)
-
-        # Initialize state
-        state = State(
-            agents=agents,
-            grid=grid,
-            pot_positions=jnp.array(self._pot_positions),
-            pot_cooking_timer=jnp.zeros(MAX_POTS, dtype=jnp.int32),
-            pot_active_mask=jnp.array(self._pot_active_mask),
-            order_types=jnp.zeros(self.max_orders, dtype=jnp.int32),
-            order_expirations=jnp.zeros(self.max_orders, dtype=jnp.int32),
-            order_active_mask=jnp.zeros(self.max_orders, dtype=jnp.bool_),
-            item_conveyor_positions=jnp.array(self._item_conveyor_positions),
-            item_conveyor_directions=jnp.array(self._item_conveyor_directions),
-            item_conveyor_active_mask=jnp.array(self._item_conveyor_active_mask),
-            player_conveyor_positions=jnp.array(self._player_conveyor_positions),
-            player_conveyor_directions=jnp.array(self._player_conveyor_directions),
-            player_conveyor_active_mask=jnp.array(self._player_conveyor_active_mask),
-            moving_wall_positions=jnp.array(self._moving_wall_positions),
-            moving_wall_directions=jnp.array(self._moving_wall_directions),
-            moving_wall_active_mask=jnp.array(self._moving_wall_active_mask),
-            moving_wall_paused=jnp.array(self._moving_wall_initial_paused),
-            moving_wall_bounce=jnp.array(self._moving_wall_bounce),
-            button_positions=jnp.array(self._button_positions),
-            button_target_idxs=jnp.array(self._button_target_idxs),
-            button_target_mask=jnp.array(self._button_target_mask),
-            button_action_type=jnp.array(self._button_action_type),
-            button_active_mask=jnp.array(self._button_active_mask),
-            button_toggled=jnp.zeros(MAX_BUTTONS, dtype=jnp.bool_),
-            barrier_positions=jnp.array(self._barrier_positions),
-            barrier_active=jnp.array(self._barrier_initial_active),
-            barrier_active_mask=jnp.array(self._barrier_active_mask),
-            barrier_timer=jnp.zeros(MAX_BARRIERS, dtype=jnp.int32),
-            barrier_duration=jnp.array(self._barrier_duration_config),
-            pressure_plate_positions=jnp.array(self._pressure_plate_positions),
-            pressure_plate_linked_barrier=jnp.array(self._pressure_plate_linked_barrier),
-            pressure_plate_action_type=jnp.array(self._pressure_plate_action_type),
-            pressure_plate_active_mask=jnp.array(self._pressure_plate_active_mask),
-            pressure_plate_toggled=jnp.zeros(MAX_PRESSURE_PLATES, dtype=jnp.bool_),
-            time=jnp.array(0),
-            terminal=False,
-            recipe=recipe,
-            new_correct_delivery=False,
-        )
-
-        # Optional random initialization
-        key, key_randomize = jax.random.split(key)
-        if self.random_reset:
-            state = self._randomize_state(state, key_randomize)
-        elif self.random_agent_positions:
-            state = self._randomize_agent_positions(state, key_randomize)
-
-        obs = self.get_obs(state)
-
-        return lax.stop_gradient(obs), lax.stop_gradient(state)
-
-    def _sample_recipe(self, key: chex.PRNGKey) -> int:
-        """Sample a recipe from possible recipes."""
-        recipe_idx = jax.random.randint(key, (), 0, self.possible_recipes.shape[0])
-        recipe = self.possible_recipes[recipe_idx]
-        return DynamicObject.get_recipe_encoding(recipe)
-
-    @staticmethod
-    def _is_agent_walkable(static_object, pos, state):
-        # Check if destination is a barrier and if it's active
-        is_barrier_tile = static_object == StaticObject.BARRIER
-        barrier_blocks = False
-        # Check all barriers to see if any active barrier is at pos
-
-        at_barrier_pos = (
-            (state.barrier_positions[:, 0] == pos.y)
-            & (state.barrier_positions[:, 1] == pos.x)
-            & state.barrier_active_mask[:]
-        )
-
-        barrier_blocks = jnp.any(at_barrier_pos & state.barrier_active)
-
-        is_walkable = (
-            (static_object == StaticObject.EMPTY)
-            | (static_object == StaticObject.PLAYER_CONVEYOR)
-            | (static_object == StaticObject.PRESSURE_PLATE)
-            | (
-                is_barrier_tile & ~barrier_blocks
-            )  # Barrier is walkable if inactive
-        )
-
-        return is_walkable
-
-    @staticmethod
-    def _barriers_occupied(agent_ys, agent_xs, barrier_positions, barrier_active_mask):
-        """Boolean [MAX_BARRIERS]: True where an agent stands on a barrier's tile.
-
-        Used to keep a barrier from (re)closing on top of an agent, so an agent
-        can never be trapped inside a barrier that activates on them.
-        """
-        on_barrier = (
-            (agent_ys[None, :] == barrier_positions[:, 0][:, None])
-            & (agent_xs[None, :] == barrier_positions[:, 1][:, None])
-        )
-        return jnp.any(on_barrier, axis=1) & barrier_active_mask
-
-    def _randomize_agent_positions(self, state: State, key: chex.PRNGKey) -> State:
-        """Randomize agent positions within their rooms."""
-        num_agents = self.num_agents
-        agents = state.agents
-
-        def _select_agent_position(taken_mask, x):
-            pos, key = x
-
-            allowed_positions = (
-                self.enclosed_spaces == self.enclosed_spaces[pos.y, pos.x]
-            ) & ~taken_mask
-            allowed_positions = allowed_positions.flatten()
-
-            p = allowed_positions / jnp.sum(allowed_positions)
-            agent_pos_idx = jax.random.choice(key, allowed_positions.size, (), p=p)
-            agent_position = Position(
-                x=agent_pos_idx % self.width, y=agent_pos_idx // self.width
-            )
-
-            new_taken_mask = taken_mask.at[agent_position.y, agent_position.x].set(True)
-            return new_taken_mask, agent_position
-
-        taken_mask = jnp.zeros_like(self.enclosed_spaces, dtype=jnp.bool_)
-        key, subkey = jax.random.split(key)
-        keys = jax.random.split(subkey, num_agents)
-        _, agent_positions = jax.lax.scan(
-            _select_agent_position, taken_mask, (agents.pos, keys)
-        )
-
-        key, subkey = jax.random.split(key)
-        directions = jax.random.randint(subkey, (num_agents,), 0, len(Direction))
-
-        return state.replace(agents=agents.replace(pos=agent_positions, dir=directions))
-
-    def _randomize_state(self, state: State, key: chex.PRNGKey) -> State:
-        """Randomize the full state."""
-        key, subkey = jax.random.split(key)
-        state = self._randomize_agent_positions(state, subkey)
-        # Could add more randomization here (pot contents, items on counters, etc.)
-        return state
-
-    # -------------------------------------------------------------------------
-    # Environment Step Pipeline
-    # -------------------------------------------------------------------------
+        """Reset the environment by running the functional reset pipeline."""
+        return reset_overcooked_v3(key, self.config)
 
     def step_env(
         self,
@@ -744,65 +473,8 @@ class OvercookedV3(MultiAgentEnv):
         state: State,
         actions: Dict[str, chex.Array],
     ) -> Tuple[Dict[str, chex.Array], State, Dict[str, float], Dict[str, bool], Dict]:
-        """Perform a single timestep state transition."""
-
-        acts = self.action_set.take(
-            indices=jnp.array([actions[f"agent_{i}"] for i in range(self.num_agents)])
-        )
-
-        state, reward, shaped_rewards = self.step_agents(key, state, acts)
-
-        # Process moving walls (before conveyors so conveyors interact with new wall positions)
-        if self.enable_moving_walls:
-            state = self._process_moving_walls(state)
-
-        if self.enable_pressure_plates:
-            state = self._process_pressure_plates(state)
-
-        # Process conveyors
-        if self.enable_item_conveyors:
-            state = self._process_item_conveyors(state)
-        if self.enable_player_conveyors:
-            state = self._process_player_conveyors(state)
-
-        # Process barrier timers (for timed barriers)
-        state = self._process_barrier_timers(state)
-
-        # Process order queue
-        if self.enable_order_queue:
-            key, subkey = jax.random.split(key)
-            state, order_reward = self._process_order_queue(state, subkey)
-            reward = reward + order_reward
-
-        # Update time
-        state = state.replace(time=state.time + 1)
-
-        # Check termination
-        done = self.is_terminal(state)
-        state = state.replace(terminal=done)
-
-        obs = self.get_obs(state)
-
-        rewards = {f"agent_{i}": reward for i in range(self.num_agents)}
-        shaped_rewards_dict = {
-            f"agent_{i}": shaped_reward
-            for i, shaped_reward in enumerate(shaped_rewards)
-        }
-
-        dones = {f"agent_{i}": done for i in range(self.num_agents)}
-        dones["__all__"] = done
-
-        return (
-            lax.stop_gradient(obs),
-            lax.stop_gradient(state),
-            rewards,
-            dones,
-            {"shaped_reward": shaped_rewards_dict},
-        )
-
-    # -------------------------------------------------------------------------
-    # Agent Movement, Collisions, and Button Interactions
-    # -------------------------------------------------------------------------
+        """Perform a single timestep by forwarding to the functional step pipeline."""
+        return self._step_jit(key, state, actions, self.config)
 
     def step_agents(
         self,
@@ -810,434 +482,8 @@ class OvercookedV3(MultiAgentEnv):
         state: State,
         actions: chex.Array,
     ) -> Tuple[State, float, chex.Array]:
-        """Process agent actions and update state."""
-        grid = state.grid
-
-        # Determine which barriers are made walkable by pressure plates.
-        # Only runs when pressure plates are enabled; otherwise all False.
-        barrier_walkable_by_pressure_plate = jnp.zeros(MAX_BARRIERS, dtype=jnp.bool_)
-
-        if self.enable_pressure_plates:
-            def _check_pressure_plate(barrier_walkable, plate_idx):
-                plate_valid = state.pressure_plate_active_mask[plate_idx]
-                linked_barrier_mask = state.pressure_plate_linked_barrier[plate_idx]
-
-                def _agent_on_plate(agent_pos):
-                    return (
-                        (agent_pos.y == state.pressure_plate_positions[plate_idx, 0])
-                        & (agent_pos.x == state.pressure_plate_positions[plate_idx, 1])
-                    )
-
-                agent_on_plate = jax.vmap(_agent_on_plate)(state.agents.pos)
-                plate_pressed = plate_valid & jnp.any(agent_on_plate)
-                updated_barrier_walkable = barrier_walkable | (linked_barrier_mask & plate_pressed)
-                return updated_barrier_walkable, None
-
-            barrier_walkable_by_pressure_plate, _ = jax.lax.scan(
-                _check_pressure_plate,
-                barrier_walkable_by_pressure_plate,
-                jnp.arange(MAX_PRESSURE_PLATES),
-            )
-
-        # Movement phase
-        def _move_wrapper(agent, action):
-            direction = ACTION_TO_DIRECTION[action]
-
-            def _move(agent, dir):
-                pos = agent.pos
-                new_pos = pos.move_in_bounds(dir, self.width, self.height)
-
-                # Check if new position is walkable
-                new_cell_static = grid[new_pos.y, new_pos.x, 0]
-
-                # Check if destination is a barrier and if it's active.
-                # Vectorized over all barriers (mirrors _is_agent_walkable) while
-                # also honoring pressure-plate walkability overrides.
-                is_barrier_tile = new_cell_static == StaticObject.BARRIER
-                at_barrier_pos = (
-                    (state.barrier_positions[:, 0] == new_pos.y)
-                    & (state.barrier_positions[:, 1] == new_pos.x)
-                    & state.barrier_active_mask
-                )
-                barrier_blocks = jnp.any(
-                    at_barrier_pos
-                    & state.barrier_active
-                    & ~barrier_walkable_by_pressure_plate
-                )
-
-                is_walkable = (
-                    (new_cell_static == StaticObject.EMPTY)
-                    | (new_cell_static == StaticObject.PLAYER_CONVEYOR)
-                    | (new_cell_static == StaticObject.PRESSURE_PLATE)
-                    | (is_barrier_tile & ~barrier_blocks)
-                )
-
-                new_pos = tree_select(is_walkable, new_pos, pos)
-                return agent.replace(pos=new_pos, dir=direction)
-
-            return jax.lax.cond(
-                direction != -1,
-                _move,
-                lambda a, _: a,
-                agent,
-                direction,
-            )
-
-        new_agents = jax.vmap(_move_wrapper)(state.agents, actions)
-
-        # Resolve collisions
-        def _masked_positions(mask):
-            return tree_select(mask, state.agents.pos, new_agents.pos)
-
-        def _get_collisions(mask):
-            positions = _masked_positions(mask)
-
-            collision_grid = jnp.zeros((self.height, self.width))
-            collision_grid, _ = jax.lax.scan(
-                lambda grid, pos: (grid.at[pos.y, pos.x].add(1), None),
-                collision_grid,
-                positions,
-            )
-
-            collision_mask = collision_grid > 1
-            collisions = jax.vmap(lambda p: collision_mask[p.y, p.x])(positions)
-            return collisions
-
-        initial_mask = jnp.zeros((self.num_agents,), dtype=bool)
-        mask = jax.lax.while_loop(
-            lambda mask: jnp.any(_get_collisions(mask)),
-            lambda mask: mask | _get_collisions(mask),
-            initial_mask,
-        )
-        new_agents = new_agents.replace(pos=_masked_positions(mask))
-
-        # Prevent swapping
-        def _compute_swapped_agents(original_positions, new_positions):
-            original_positions = original_positions.to_array()
-            new_positions = new_positions.to_array()
-
-            original_pos_expanded = jnp.expand_dims(original_positions, axis=0)
-            new_pos_expanded = jnp.expand_dims(new_positions, axis=1)
-
-            swap_mask = (original_pos_expanded == new_pos_expanded).all(axis=-1)
-            swap_mask = jnp.fill_diagonal(swap_mask, False, inplace=False)
-
-            swap_pairs = jnp.logical_and(swap_mask, swap_mask.T)
-            swapped_agents = jnp.any(swap_pairs, axis=0)
-            return swapped_agents
-
-        swap_mask = _compute_swapped_agents(state.agents.pos, new_agents.pos)
-        new_agents = new_agents.replace(pos=_masked_positions(swap_mask))
-
-        # Interaction phase
-        def _interact_wrapper(carry, x):
-            agent, action = x
-            is_interact = action == Actions.interact
-
-            def _interact(carry, agent):
-                grid, correct_delivery, reward, pot_timers = carry
-
-                (
-                    new_grid,
-                    new_agent,
-                    new_correct_delivery,
-                    interact_reward,
-                    shaped_reward,
-                    new_pot_timers,
-                ) = self.process_interact(
-                    grid,
-                    agent,
-                    new_agents.inventory,
-                    state.recipe,
-                    pot_timers,
-                    state.pot_positions,
-                    state.pot_active_mask,
-                )
-
-                carry = (
-                    new_grid,
-                    correct_delivery | new_correct_delivery,
-                    reward + interact_reward,
-                    new_pot_timers,
-                )
-                return carry, (new_agent, shaped_reward)
-
-            return jax.lax.cond(
-                is_interact, _interact, lambda c, a: (c, (a, 0.0)), carry, agent
-            )
-
-        carry = (grid, False, 0.0, state.pot_cooking_timer)
-        xs = (new_agents, actions)
-        (
-            (new_grid, new_correct_delivery, reward, new_pot_timers),
-            (new_agents, shaped_rewards),
-        ) = jax.lax.scan(_interact_wrapper, carry, xs)
-
-        # Update pot timers (cooking and burning)
-        new_grid, new_pot_timers = self._update_pot_timers(
-            new_grid, new_pot_timers, state.pot_positions, state.pot_active_mask
-        )
-
-        # Process button presses
-        new_mw_directions = state.moving_wall_directions
-        new_mw_paused = state.moving_wall_paused
-        new_mw_bounce = state.moving_wall_bounce
-        new_btn_toggled = state.button_toggled
-        new_barrier_active = state.barrier_active
-        new_barrier_timer = state.barrier_timer
-
-        if self.enable_buttons:
-            # Barriers a toggle button must not close because an agent stands on
-            # them (post-movement positions); the close is skipped until clear.
-            barrier_occupied = self._barriers_occupied(
-                new_agents.pos.y,
-                new_agents.pos.x,
-                state.barrier_positions,
-                state.barrier_active_mask,
-            )
-
-            def _process_agent_button(carry, x):
-                mw_dirs, mw_paused, mw_bounce, btn_toggled, bar_active, bar_timer = (
-                    carry
-                )
-                agent, action = x
-                is_interact = action == Actions.interact
-                fwd_pos = agent.get_fwd_pos()
-                fwd_static = new_grid[fwd_pos.y, fwd_pos.x, 0]
-                is_button = fwd_static == StaticObject.BUTTON
-
-                def _scan_buttons(carry):
-                    (
-                        mw_dirs,
-                        mw_paused,
-                        mw_bounce,
-                        btn_toggled,
-                        bar_active,
-                        bar_timer,
-                    ) = carry
-
-                    def _check_button(carry, button_idx):
-                        (
-                            mw_dirs,
-                            mw_paused,
-                            mw_bounce,
-                            btn_toggled,
-                            bar_active,
-                            bar_timer,
-                        ) = carry
-                        btn_y = state.button_positions[button_idx, 0]
-                        btn_x = state.button_positions[button_idx, 1]
-                        is_active = state.button_active_mask[button_idx]
-                        is_this = (
-                            (btn_y == fwd_pos.y) & (btn_x == fwd_pos.x) & is_active
-                        )
-
-                        action_type = state.button_action_type[button_idx]
-
-                        # Toggle button state
-                        new_toggled = jax.lax.select(
-                            is_this, ~btn_toggled[button_idx], btn_toggled[button_idx]
-                        )
-                        btn_toggled = btn_toggled.at[button_idx].set(new_toggled)
-
-                        def _apply_target(carry, target_slot):
-                            (
-                                mw_dirs,
-                                mw_paused,
-                                mw_bounce,
-                                bar_active,
-                                bar_timer,
-                            ) = carry
-                            target_idx = state.button_target_idxs[
-                                button_idx, target_slot
-                            ]
-                            target_enabled = state.button_target_mask[
-                                button_idx, target_slot
-                            ]
-                            should_apply = is_this & target_enabled
-                            mw_idx = jnp.clip(target_idx, 0, MAX_MOVING_WALLS - 1)
-                            barrier_idx = jnp.clip(target_idx, 0, MAX_BARRIERS - 1)
-
-                            # Moving wall actions
-                            mw_paused = jax.lax.select(
-                                should_apply
-                                & (action_type == ButtonAction.TOGGLE_PAUSE),
-                                mw_paused.at[mw_idx].set(~mw_paused[mw_idx]),
-                                mw_paused,
-                            )
-
-                            new_dir = Direction.opposite(mw_dirs[mw_idx])
-                            mw_dirs = jax.lax.select(
-                                should_apply
-                                & (action_type == ButtonAction.TOGGLE_DIRECTION),
-                                mw_dirs.at[mw_idx].set(new_dir),
-                                mw_dirs,
-                            )
-
-                            mw_bounce = jax.lax.select(
-                                should_apply
-                                & (action_type == ButtonAction.TOGGLE_BOUNCE),
-                                mw_bounce.at[mw_idx].set(~mw_bounce[mw_idx]),
-                                mw_bounce,
-                            )
-
-                            mw_paused = jax.lax.select(
-                                should_apply
-                                & (action_type == ButtonAction.TRIGGER_MOVE),
-                                mw_paused.at[mw_idx].set(False),
-                                mw_paused,
-                            )
-
-                            # Barrier actions. A toggle never closes a barrier on
-                            # top of an agent; the close is skipped (left open) so
-                            # a later press can close it once the tile is clear.
-                            toggled_active = ~bar_active[barrier_idx]
-                            safe_active = jnp.where(
-                                toggled_active & barrier_occupied[barrier_idx],
-                                bar_active[barrier_idx],
-                                toggled_active,
-                            )
-                            bar_active = jax.lax.select(
-                                should_apply
-                                & (action_type == ButtonAction.TOGGLE_BARRIER),
-                                bar_active.at[barrier_idx].set(safe_active),
-                                bar_active,
-                            )
-
-                            bar_active = jax.lax.select(
-                                should_apply
-                                & (action_type == ButtonAction.TIMED_BARRIER),
-                                bar_active.at[barrier_idx].set(False),
-                                bar_active,
-                            )
-                            bar_timer = jax.lax.select(
-                                should_apply
-                                & (action_type == ButtonAction.TIMED_BARRIER),
-                                bar_timer.at[barrier_idx].set(
-                                    state.barrier_duration[barrier_idx]
-                                ),
-                                bar_timer,
-                            )
-
-                            return (
-                                mw_dirs,
-                                mw_paused,
-                                mw_bounce,
-                                bar_active,
-                                bar_timer,
-                            ), None
-
-                        (
-                            mw_dirs,
-                            mw_paused,
-                            mw_bounce,
-                            bar_active,
-                            bar_timer,
-                        ), _ = jax.lax.scan(
-                            _apply_target,
-                            (
-                                mw_dirs,
-                                mw_paused,
-                                mw_bounce,
-                                bar_active,
-                                bar_timer,
-                            ),
-                            jnp.arange(MAX_BUTTON_TARGETS),
-                        )
-
-                        return (
-                            mw_dirs,
-                            mw_paused,
-                            mw_bounce,
-                            btn_toggled,
-                            bar_active,
-                            bar_timer,
-                        ), None
-
-                    (
-                        (
-                            mw_dirs,
-                            mw_paused,
-                            mw_bounce,
-                            btn_toggled,
-                            bar_active,
-                            bar_timer,
-                        ),
-                        _,
-                    ) = jax.lax.scan(
-                        _check_button,
-                        (
-                            mw_dirs,
-                            mw_paused,
-                            mw_bounce,
-                            btn_toggled,
-                            bar_active,
-                            bar_timer,
-                        ),
-                        jnp.arange(MAX_BUTTONS),
-                    )
-                    return (
-                        mw_dirs,
-                        mw_paused,
-                        mw_bounce,
-                        btn_toggled,
-                        bar_active,
-                        bar_timer,
-                    )
-
-                should_process = is_interact & is_button
-                new_carry = jax.lax.cond(
-                    should_process,
-                    _scan_buttons,
-                    lambda c: c,
-                    (mw_dirs, mw_paused, mw_bounce, btn_toggled, bar_active, bar_timer),
-                )
-
-                return new_carry, None
-
-            (
-                (
-                    new_mw_directions,
-                    new_mw_paused,
-                    new_mw_bounce,
-                    new_btn_toggled,
-                    new_barrier_active,
-                    new_barrier_timer,
-                ),
-                _,
-            ) = jax.lax.scan(
-                _process_agent_button,
-                (
-                    new_mw_directions,
-                    new_mw_paused,
-                    new_mw_bounce,
-                    new_btn_toggled,
-                    new_barrier_active,
-                    new_barrier_timer,
-                ),
-                (new_agents, actions),
-            )
-
-        return (
-            state.replace(
-                agents=new_agents,
-                grid=new_grid,
-                pot_cooking_timer=new_pot_timers,
-                new_correct_delivery=new_correct_delivery,
-                moving_wall_directions=new_mw_directions,
-                moving_wall_paused=new_mw_paused,
-                moving_wall_bounce=new_mw_bounce,
-                button_toggled=new_btn_toggled,
-                barrier_active=new_barrier_active,
-                barrier_timer=new_barrier_timer,
-            ),
-            reward,
-            shaped_rewards,
-        )
-
-    # -------------------------------------------------------------------------
-    # Interact Action Handling
-    # -------------------------------------------------------------------------
+        """Compatibility wrapper for the functional agent action phase."""
+        return run_agent_action_phase(key, state, actions, self.config)
 
     def process_interact(
         self,
@@ -1249,202 +495,51 @@ class OvercookedV3(MultiAgentEnv):
         pot_positions: chex.Array,
         pot_active_mask: chex.Array,
     ):
-        """Process an interact action for an agent."""
-        inventory = agent.inventory
-        fwd_pos, fwd_pos_in_bounds = agent.pos.checked_move(
-            agent.dir, self.width, self.height
+        """Compatibility wrapper for functional interact processing."""
+        return process_interact(
+            grid,
+            agent,
+            all_inventories,
+            recipe,
+            pot_timers,
+            pot_positions,
+            pot_active_mask,
+            self.config,
         )
 
-        shaped_reward = jnp.array(0.0, dtype=float)
-
-        interact_cell = grid[fwd_pos.y, fwd_pos.x]
-        interact_item = interact_cell[0]
-        interact_ingredients = interact_cell[1]
-        interact_extra = interact_cell[2]
-
-        plated_recipe = recipe | DynamicObject.PLATE | DynamicObject.COOKED
-
-        # What is the object?
-        object_is_plate_pile = fwd_pos_in_bounds & (
-            interact_item == StaticObject.PLATE_PILE
-        )
-        object_is_ingredient_pile = (
-            fwd_pos_in_bounds & StaticObject.is_ingredient_pile(interact_item)
-        )
-        object_is_pile = object_is_plate_pile | object_is_ingredient_pile
-
-        object_is_pot = fwd_pos_in_bounds & (interact_item == StaticObject.POT)
-        object_is_goal = fwd_pos_in_bounds & (interact_item == StaticObject.GOAL)
-        object_is_wall = fwd_pos_in_bounds & (
-            (interact_item == StaticObject.WALL)
-            | (interact_item == StaticObject.MOVING_WALL)
-        )
-        object_is_conveyor = fwd_pos_in_bounds & (
-            (interact_item == StaticObject.ITEM_CONVEYOR)
-            | (interact_item == StaticObject.PLAYER_CONVEYOR)
-        )
-        object_has_no_ingredients = interact_ingredients == 0
-
-        # What is in inventory?
-        inventory_is_empty = inventory == 0
-        inventory_is_ingredient = DynamicObject.is_ingredient(inventory)
-        inventory_is_plate = inventory == DynamicObject.PLATE
-        inventory_is_dish = (inventory & DynamicObject.COOKED) != 0
-
-        merged_ingredients = interact_ingredients + inventory
-
-        # Pot state
-        pot_is_cooked = object_is_pot * (
-            (interact_ingredients & DynamicObject.COOKED) != 0
-        )
-        pot_is_cooking = object_is_pot * (interact_extra > 0) * ~pot_is_cooked
-        pot_is_burned = object_is_pot * (
-            (interact_ingredients & DynamicObject.BURNED) != 0
-        )
-        pot_is_idle = object_is_pot * ~pot_is_cooking * ~pot_is_cooked * ~pot_is_burned
-
-        # Check if pot is ready (in burning window)
-        # In V3: dish_ready when cooking_timer is between 1 and burn_time
-        pot_is_ready = pot_is_cooked
-
-        # Pickup success conditions
-        successful_dish_pickup = pot_is_ready * inventory_is_plate
-        is_dish_pickup_useful = merged_ingredients == plated_recipe
-        if self.shaped_rewards_enabled:
-            shaped_reward += (
-                successful_dish_pickup
-                * is_dish_pickup_useful
-                * SHAPED_REWARDS["SOUP_IN_DISH"]
-            )
-
-        successful_pickup = (
-            object_is_pile * inventory_is_empty
-            + successful_dish_pickup
-            + object_is_wall * ~object_has_no_ingredients * inventory_is_empty
-            + object_is_conveyor * ~object_has_no_ingredients * inventory_is_empty
+    def _get_obs_shape(self) -> Tuple[int, ...]:
+        """Return the configured observation shape."""
+        return calculate_observation_shape(
+            self.width,
+            self.height,
+            self.layout,
+            self.observation_type,
+            self.agent_view_size,
         )
 
-        # Pot placement
-        pot_full = DynamicObject.ingredient_count(interact_ingredients) == 3
+    def _sample_recipe(self, key: chex.PRNGKey) -> int:
+        """Compatibility wrapper for functional recipe sampling."""
+        return sample_recipe(key, self.config)
 
-        # Check same ingredient type for pot
-        pot_ingredient_type = DynamicObject.get_ingredient_type(interact_ingredients)
-        inventory_ingredient_type = DynamicObject.get_ingredient_type(inventory)
-        same_ingredient_type = (pot_ingredient_type == inventory_ingredient_type) | (
-            interact_ingredients == 0
+    def _randomize_agent_positions(self, state: State, key: chex.PRNGKey) -> State:
+        """Compatibility wrapper for functional agent position randomization."""
+        return randomize_agent_positions(state, key, self.config)
+
+    def _randomize_state(self, state: State, key: chex.PRNGKey) -> State:
+        """Compatibility wrapper for functional state randomization."""
+        return randomize_state(state, key, self.config)
+
+    @staticmethod
+    def _is_agent_walkable(static_object, pos, state):
+        """Compatibility wrapper for functional walkability checks."""
+        return is_agent_walkable(static_object, pos, state)
+
+    @staticmethod
+    def _barriers_occupied(agent_ys, agent_xs, barrier_positions, barrier_active_mask):
+        """Compatibility wrapper for functional barrier occupancy checks."""
+        return barriers_occupied(
+            agent_ys, agent_xs, barrier_positions, barrier_active_mask
         )
-
-        successful_pot_placement = (
-            pot_is_idle * inventory_is_ingredient * ~pot_full * same_ingredient_type
-        )
-        ingredient_selector = inventory | (inventory << 1)
-        is_pot_placement_useful = (interact_ingredients & ingredient_selector) < (
-            recipe & ingredient_selector
-        )
-        if self.shaped_rewards_enabled:
-            shaped_reward += (
-                successful_pot_placement
-                * is_pot_placement_useful
-                * SHAPED_REWARDS["PLACEMENT_IN_POT"]
-            )
-
-        # Drop on counter/conveyor
-        successful_drop = (
-            object_is_wall | object_is_conveyor
-        ) * object_has_no_ingredients * ~inventory_is_empty + successful_pot_placement
-
-        # Delivery
-        successful_delivery = object_is_goal * inventory_is_dish
-        no_effect = ~successful_pickup * ~successful_drop * ~successful_delivery
-
-        # Compute new ingredient layer
-        pile_ingredient = (
-            object_is_plate_pile * DynamicObject.PLATE
-            + object_is_ingredient_pile * StaticObject.get_ingredient(interact_item)
-        )
-
-        new_ingredients = (
-            successful_drop * merged_ingredients + no_effect * interact_ingredients
-        )
-
-        # Start cooking when pot becomes full
-        pot_full_after_drop = DynamicObject.ingredient_count(new_ingredients) == 3
-        auto_cook = pot_is_idle & pot_full_after_drop
-
-        # Update pot timer
-        # Find which pot this is
-        def _update_pot_timer(pot_idx):
-            pot_y, pot_x = pot_positions[pot_idx]
-            is_this_pot = (
-                (pot_y == fwd_pos.y) & (pot_x == fwd_pos.x) & pot_active_mask[pot_idx]
-            )
-            new_timer = jax.lax.select(
-                is_this_pot & auto_cook, self.pot_cook_time, pot_timers[pot_idx]
-            )
-            # Reset timer on successful dish pickup
-            new_timer = jax.lax.select(
-                is_this_pot & successful_dish_pickup, 0, new_timer
-            )
-            return new_timer
-
-        new_pot_timers = jax.vmap(_update_pot_timer)(jnp.arange(MAX_POTS))
-
-        new_extra = interact_extra  # Keep conveyor directions etc
-
-        new_cell = jnp.array([interact_item, new_ingredients, new_extra])
-        new_grid = grid.at[fwd_pos.y, fwd_pos.x].set(new_cell)
-
-        new_inventory = (
-            successful_pickup * (pile_ingredient + merged_ingredients)
-            + no_effect * inventory
-        )
-        new_agent = agent.replace(inventory=new_inventory)
-
-        # Reward calculation
-        is_correct_recipe = inventory == plated_recipe
-
-        reward = jnp.array(0.0, dtype=float)
-        reward += (
-            successful_delivery
-            * jax.lax.select(is_correct_recipe, 1.0, 0.0)
-            * self.delivery_reward
-        )
-
-        # Plate pickup reward
-        if self.shaped_rewards_enabled:
-            inventory_is_plate_now = new_inventory == DynamicObject.PLATE
-            successful_plate_pickup = successful_pickup * inventory_is_plate_now
-            num_plates_in_inventory = jnp.sum(all_inventories == DynamicObject.PLATE)
-            pot_ingredient_counts = jax.vmap(jax.vmap(DynamicObject.ingredient_count))(
-                grid[:, :, 1]
-            )
-            full_unburned_pots = (
-                (grid[:, :, 0] == StaticObject.POT)
-                & (pot_ingredient_counts == 3)
-                & ((grid[:, :, 1] & DynamicObject.BURNED) == 0)
-            )
-            num_useful_pots = jnp.sum(full_unburned_pots)
-            is_plate_pickup_useful = num_plates_in_inventory < num_useful_pots
-            shaped_reward += (
-                is_plate_pickup_useful
-                * successful_plate_pickup
-                * SHAPED_REWARDS["PLATE_PICKUP"]
-            )
-
-        correct_delivery = successful_delivery & is_correct_recipe
-
-        return (
-            new_grid,
-            new_agent,
-            correct_delivery,
-            reward,
-            shaped_reward,
-            new_pot_timers,
-        )
-
-    # -------------------------------------------------------------------------
-    # Dynamic Environment Systems
-    # -------------------------------------------------------------------------
 
     def _update_pot_timers(
         self,
@@ -1453,713 +548,54 @@ class OvercookedV3(MultiAgentEnv):
         pot_positions: chex.Array,
         pot_active_mask: chex.Array,
     ) -> Tuple[chex.Array, chex.Array]:
-        """Update pot cooking timers and handle burning."""
-
-        def _update_single_pot(carry, pot_idx):
-            grid, timers = carry
-            pot_y, pot_x = pot_positions[pot_idx]
-            is_active = pot_active_mask[pot_idx]
-            current_timer = timers[pot_idx]
-
-            pot_cell = grid[pot_y, pot_x]
-            pot_ingredients = pot_cell[1]
-
-            # Check if pot is full (has 3 ingredients)
-            ingredient_count = DynamicObject.ingredient_count(pot_ingredients)
-            pot_is_full = ingredient_count == 3
-            pot_is_cooking = (current_timer > 0) & pot_is_full
-            pot_already_cooked = (pot_ingredients & DynamicObject.COOKED) != 0
-
-            # Decrement timer if cooking
-            new_timer = jax.lax.select(
-                is_active & pot_is_cooking, current_timer - 1, current_timer
-            )
-
-            # Check if just finished cooking (entered burning window)
-            just_finished_cooking = pot_is_cooking & (new_timer == self.pot_burn_time)
-            # Mark as cooked when timer reaches burn_time
-            new_ingredients = jax.lax.select(
-                is_active & just_finished_cooking,
-                pot_ingredients | DynamicObject.COOKED,
-                pot_ingredients,
-            )
-
-            # Check if pot burned (timer hit 0 while cooking)
-            just_burned = pot_is_cooking & (new_timer == 0)
-            # Reset pot if burned
-            new_ingredients = jax.lax.select(
-                is_active & just_burned,
-                jnp.int32(0),  # Clear pot
-                new_ingredients,
-            )
-            new_timer = jax.lax.select(is_active & just_burned, jnp.int32(0), new_timer)
-
-            # Update grid
-            new_cell = jnp.array([pot_cell[0], new_ingredients, pot_cell[2]])
-            new_grid = grid.at[pot_y, pot_x].set(new_cell)
-
-            # Update timers
-            new_timers = timers.at[pot_idx].set(new_timer)
-
-            return (new_grid, new_timers), None
-
-        (new_grid, new_timers), _ = jax.lax.scan(
-            _update_single_pot, (grid, pot_timers), jnp.arange(MAX_POTS)
+        """Compatibility wrapper for functional pot timer updates."""
+        return update_pot_timers(
+            grid, pot_timers, pot_positions, pot_active_mask, self.config
         )
-
-        return new_grid, new_timers
 
     def _process_item_conveyors(self, state: State) -> State:
-        """Move items on item conveyor belts."""
-        if not self.enable_item_conveyors:
-            return state
-
-        grid = state.grid
-
-        def _move_item_on_conveyor(grid, conveyor_idx):
-            pos = state.item_conveyor_positions[conveyor_idx]
-            direction = state.item_conveyor_directions[conveyor_idx]
-            is_active = state.item_conveyor_active_mask[conveyor_idx]
-
-            y, x = pos[0], pos[1]
-            current_item = grid[y, x, 1]
-            has_item = current_item != 0
-
-            # Calculate destination
-            dir_vec = DIR_TO_VEC[direction]
-
-            raw_dest_x = x + dir_vec[0]
-            raw_dest_y = y + dir_vec[1]
-            dest_in_bounds = (
-                (raw_dest_x >= 0)
-                & (raw_dest_x < self.width)
-                & (raw_dest_y >= 0)
-                & (raw_dest_y < self.height)
-            )
-            dest_x = jnp.clip(raw_dest_x, 0, self.width - 1)
-            dest_y = jnp.clip(raw_dest_y, 0, self.height - 1)
-
-            # Check if destination can receive item
-            dest_static = grid[dest_y, dest_x, 0]
-            dest_item = grid[dest_y, dest_x, 1]
-            dest_can_receive = (
-                dest_in_bounds
-                & (
-                    (dest_static == StaticObject.WALL)
-                    | (dest_static == StaticObject.ITEM_CONVEYOR)
-                    | (dest_static == StaticObject.PLAYER_CONVEYOR)
-                    | (dest_static == StaticObject.GOAL)
-                    | (dest_static == StaticObject.MOVING_WALL)
-                )
-                & (dest_item == 0)
-            )
-
-            should_move = is_active & has_item & dest_can_receive
-            should_disappear = is_active & has_item & ~dest_in_bounds
-
-            # Move item
-            new_grid = jax.lax.select(
-                should_disappear,
-                grid.at[y, x, 1].set(0),
-                jax.lax.select(
-                    should_move,
-                    grid.at[y, x, 1].set(0).at[dest_y, dest_x, 1].set(current_item),
-                    grid,
-                )
-            )
-
-            return new_grid, None
-
-        new_grid, _ = jax.lax.scan(
-            _move_item_on_conveyor, grid, jnp.arange(MAX_ITEM_CONVEYORS)
-        )
-
-        return state.replace(grid=new_grid)
+        """Compatibility wrapper for functional item conveyor movement."""
+        return move_items_on_item_conveyors(state, self.config)
 
     def _process_player_conveyors(self, state: State) -> State:
-        """Push agents on player conveyor belts."""
-        if not self.enable_player_conveyors:
-            return state
-
-        agents = state.agents
-        grid = state.grid
-
-        def _check_agent_on_conveyor(agent_pos, conveyor_idx):
-            pos = state.player_conveyor_positions[conveyor_idx]
-            is_active = state.player_conveyor_active_mask[conveyor_idx]
-            is_on = (agent_pos.x == pos[1]) & (agent_pos.y == pos[0]) & is_active
-            return is_on, state.player_conveyor_directions[conveyor_idx]
-
-        def _push_agent(agent):
-            # Check all conveyors
-            on_conveyor_checks = jax.vmap(
-                lambda idx: _check_agent_on_conveyor(agent.pos, idx)
-            )(jnp.arange(MAX_PLAYER_CONVEYORS))
-
-            is_on_any, directions = on_conveyor_checks
-            # Take first active conveyor's direction
-            conveyor_idx = jnp.argmax(is_on_any)
-            is_on = jnp.any(is_on_any)
-            push_direction = directions[conveyor_idx]
-
-            # Calculate new position
-            new_pos = agent.pos.move_in_bounds(push_direction, self.width, self.height)
-
-            # Check if destination is walkable
-            dest_static = grid[new_pos.y, new_pos.x, 0]
-            dest_walkable = self._is_agent_walkable(dest_static, new_pos, state)
-
-            should_push = is_on & dest_walkable
-
-            final_pos = tree_select(should_push, new_pos, agent.pos)
-            return agent.replace(pos=final_pos)
-
-        new_agents = jax.vmap(_push_agent)(agents)
-
-        return state.replace(agents=new_agents)
+        """Compatibility wrapper for functional player conveyor movement."""
+        return push_players_on_player_conveyors(state, self.config)
 
     def _process_barrier_timers(self, state: State) -> State:
-        """Decrement barrier timers and reactivate barriers when timers reach zero."""
-
-        # A barrier is never reactivated on top of an agent; the timer is held
-        # until the tile clears so the agent can't be trapped (see
-        # _barriers_occupied).
-        barrier_occupied = self._barriers_occupied(
-            state.agents.pos.y,
-            state.agents.pos.x,
-            state.barrier_positions,
-            state.barrier_active_mask,
-        )
-
-        def _update_single_barrier(i):
-            is_active_slot = state.barrier_active_mask[i]
-            timer = state.barrier_timer[i]
-            barrier_active = state.barrier_active[i]
-            occupied = barrier_occupied[i]
-
-            # The barrier would reactivate when its timer ticks from 1 to 0, but
-            # if an agent is on the tile we hold the timer at 1 and defer.
-            would_reactivate = timer == 1
-            hold = would_reactivate & occupied
-
-            # Timer only decrements when > 0 (and not held by an occupant)
-            has_timer = timer > 0
-            new_timer = jax.lax.select(has_timer & ~hold, timer - 1, timer)
-
-            should_reactivate = is_active_slot & would_reactivate & ~occupied
-            new_active = jax.lax.select(should_reactivate, True, barrier_active)
-
-            return new_timer, new_active
-
-        # Process all barriers
-        new_timers, new_active_states = jax.vmap(_update_single_barrier)(
-            jnp.arange(MAX_BARRIERS)
-        )
-
-        return state.replace(
-            barrier_timer=new_timers,
-            barrier_active=new_active_states,
-        )
+        """Compatibility wrapper for functional barrier timer updates."""
+        return update_barrier_timers(state, self.config)
 
     def _process_pressure_plates(self, state: State) -> State:
-        if not self.enable_pressure_plates:
-            return state
-
-        agent_xs = state.agents.pos.x  # [num_agents]
-        agent_ys = state.agents.pos.y  # [num_agents]
-
-        # Which plates currently have an agent standing on them? [num_plates]
-        plate_py = state.pressure_plate_positions[:, 0]
-        plate_px = state.pressure_plate_positions[:, 1]
-        agent_on_plate = jnp.any(
-            (agent_xs[None, :] == plate_px[:, None])
-            & (agent_ys[None, :] == plate_py[:, None]),
-            axis=1,
-        )
-        pressed = agent_on_plate & state.pressure_plate_active_mask  # [num_plates]
-
-        action = state.pressure_plate_action_type  # [num_plates]
-        is_toggle = action == ButtonAction.TOGGLE_BARRIER
-        is_timed = action == ButtonAction.TIMED_BARRIER
-
-        # Valid (plate, barrier) links. [num_plates, num_barriers]
-        linked = state.pressure_plate_linked_barrier & state.barrier_active_mask[None, :]
-
-        # A barrier is never re-closed on top of an agent; it stays open until
-        # the tile is clear, so an agent crossing as the plate releases can't be
-        # trapped (see _barriers_occupied).
-        barrier_occupied = self._barriers_occupied(
-            agent_ys, agent_xs, state.barrier_positions, state.barrier_active_mask
-        )
-
-        # TOGGLE_BARRIER: a linked barrier is open *exactly while* one of its
-        # plates is pressed (or an agent stands on it), and closes again the
-        # moment every plate is released and the tile is clear.
-        toggle_links = linked & is_toggle[:, None]
-        toggle_controlled = jnp.any(toggle_links, axis=0)  # [num_barriers]
-        toggle_open = jnp.any(toggle_links & pressed[:, None], axis=0)  # [num_barriers]
-        new_barrier_active = jnp.where(
-            toggle_controlled, ~toggle_open & ~barrier_occupied, state.barrier_active
-        )
-
-        # TIMED_BARRIER: pressing opens the barrier and (re)arms its timer; the
-        # barrier reactivates on its own in _process_barrier_timers.
-        timed_open = jnp.any(linked & is_timed[:, None] & pressed[:, None], axis=0)
-        new_barrier_active = jnp.where(timed_open, False, new_barrier_active)
-        new_barrier_timer = jnp.where(
-            timed_open, state.barrier_duration, state.barrier_timer
-        )
-
-        return state.replace(
-            barrier_active=new_barrier_active,
-            barrier_timer=new_barrier_timer,
-            pressure_plate_toggled=pressed,
-        )
+        """Compatibility wrapper for functional pressure plate updates."""
+        return update_pressure_plates(state, self.config)
 
     def _process_moving_walls(self, state: State) -> State:
-        """Move moving walls one step in their direction, pushing agents if needed."""
-        if not self.enable_moving_walls:
-            return state
-
-        grid = state.grid
-
-        # Build agent position arrays for collision checking
-        agent_xs = state.agents.pos.x  # [num_agents]
-        agent_ys = state.agents.pos.y  # [num_agents]
-
-        def _move_single_wall(carry, wall_idx):
-            grid, positions, directions, paused, bounce, ag_xs, ag_ys = carry
-
-            y = positions[wall_idx, 0]
-            x = positions[wall_idx, 1]
-            direction = directions[wall_idx]
-            is_active = state.moving_wall_active_mask[wall_idx]
-            is_paused = paused[wall_idx]
-            should_process = is_active & ~is_paused
-
-            # Direction vector
-            dir_vec = DIR_TO_VEC[direction]
-            dest_x = x + dir_vec[0]
-            dest_y = y + dir_vec[1]
-
-            # Bounds check
-            in_bounds = (
-                (dest_x >= 0)
-                & (dest_x < self.width)
-                & (dest_y >= 0)
-                & (dest_y < self.height)
-            )
-
-            # Safe indices for array access
-            safe_dest_x = jnp.clip(dest_x, 0, self.width - 1)
-            safe_dest_y = jnp.clip(dest_y, 0, self.height - 1)
-
-            # Check destination cell
-            dest_static = grid[safe_dest_y, safe_dest_x, 0]
-            dest_is_empty = dest_static == StaticObject.EMPTY
-
-            # Check if an agent is at the destination
-            agent_at_dest = (ag_xs == safe_dest_x) & (ag_ys == safe_dest_y)
-            any_agent_at_dest = jnp.any(agent_at_dest)
-
-            # If agent at dest, check if we can push them
-            # Beyond position = dest + dir_vec
-            beyond_x = safe_dest_x + dir_vec[0]
-            beyond_y = safe_dest_y + dir_vec[1]
-            beyond_in_bounds = (
-                (beyond_x >= 0)
-                & (beyond_x < self.width)
-                & (beyond_y >= 0)
-                & (beyond_y < self.height)
-            )
-            safe_beyond_x = jnp.clip(beyond_x, 0, self.width - 1)
-            safe_beyond_y = jnp.clip(beyond_y, 0, self.height - 1)
-
-            beyond_static = grid[safe_beyond_y, safe_beyond_x, 0]
-
-            # Check if beyond position has an active barrier
-            is_barrier_tile = beyond_static == StaticObject.BARRIER
-            barrier_blocks = False
-            for i in range(MAX_BARRIERS):
-                at_barrier_pos = (
-                    (state.barrier_positions[i, 0] == safe_beyond_y)
-                    & (state.barrier_positions[i, 1] == safe_beyond_x)
-                    & state.barrier_active_mask[i]
-                )
-                barrier_blocks = barrier_blocks | (
-                    at_barrier_pos & state.barrier_active[i]
-                )
-
-            beyond_walkable = (
-                (beyond_static == StaticObject.EMPTY)
-                | (beyond_static == StaticObject.ITEM_CONVEYOR)
-                | (beyond_static == StaticObject.PLAYER_CONVEYOR)
-                | (beyond_static == StaticObject.PRESSURE_PLATE)
-                | (is_barrier_tile & ~barrier_blocks)
-            )
-            # Also check no other agent at beyond position
-            agent_at_beyond = (ag_xs == safe_beyond_x) & (ag_ys == safe_beyond_y)
-            no_agent_at_beyond = ~jnp.any(agent_at_beyond)
-
-            can_push_agent = (
-                any_agent_at_dest
-                & beyond_in_bounds
-                & beyond_walkable
-                & no_agent_at_beyond
-            )
-
-            # Wall can move if dest is empty (no agent) or if we can push the agent
-            # Only allow pushing agents on empty tiles to avoid overwriting static objects
-            can_move = (
-                should_process
-                & in_bounds
-                & (
-                    (dest_is_empty & ~any_agent_at_dest)
-                    | (can_push_agent & (dest_static == StaticObject.EMPTY))
-                )
-            )
-
-            # Handle bounce: if blocked and bounce enabled, reverse direction
-            is_blocked = should_process & ~can_move
-            should_bounce = is_blocked & bounce[wall_idx]
-            new_direction = jax.lax.select(
-                should_bounce,
-                Direction.opposite(direction),
-                direction,
-            )
-
-            # Get item carried by this wall
-            old_item = grid[y, x, 1]
-
-            # Clear old position (becomes EMPTY with no item)
-            cleared_grid = jax.lax.select(
-                can_move,
-                grid.at[y, x].set(jnp.array([StaticObject.EMPTY, 0, 0])),
-                grid,
-            )
-
-            # Set new position with MOVING_WALL + carried item + direction in extra channel
-            new_cell = jnp.array([StaticObject.MOVING_WALL, old_item, new_direction])
-            moved_grid = jax.lax.select(
-                can_move,
-                cleared_grid.at[safe_dest_y, safe_dest_x].set(new_cell),
-                cleared_grid,
-            )
-
-            # Update direction in grid at current position if staying (bounce case)
-            final_grid = jax.lax.select(
-                ~can_move & should_bounce,
-                moved_grid.at[y, x, 2].set(new_direction),
-                moved_grid,
-            )
-
-            # Update position array
-            new_y = jax.lax.select(can_move, safe_dest_y, y)
-            new_x = jax.lax.select(can_move, safe_dest_x, x)
-            new_positions = positions.at[wall_idx].set(jnp.array([new_y, new_x]))
-
-            # Update direction array
-            new_directions = directions.at[wall_idx].set(new_direction)
-
-            # Push agent: update agent positions if we pushed
-            # Find which agent was pushed (first match)
-            push_agent_idx = jnp.argmax(agent_at_dest)
-            new_ag_xs = jax.lax.select(
-                can_move & can_push_agent,
-                ag_xs.at[push_agent_idx].set(safe_beyond_x),
-                ag_xs,
-            )
-            new_ag_ys = jax.lax.select(
-                can_move & can_push_agent,
-                ag_ys.at[push_agent_idx].set(safe_beyond_y),
-                ag_ys,
-            )
-
-            return (
-                final_grid,
-                new_positions,
-                new_directions,
-                paused,
-                bounce,
-                new_ag_xs,
-                new_ag_ys,
-            ), None
-
-        init_carry = (
-            grid,
-            state.moving_wall_positions,
-            state.moving_wall_directions,
-            state.moving_wall_paused,
-            state.moving_wall_bounce,
-            agent_xs,
-            agent_ys,
-        )
-
-        (
-            (
-                new_grid,
-                new_positions,
-                new_directions,
-                new_paused,
-                _bounce,
-                new_ag_xs,
-                new_ag_ys,
-            ),
-            _,
-        ) = jax.lax.scan(_move_single_wall, init_carry, jnp.arange(MAX_MOVING_WALLS))
-
-        # Re-pause walls linked to TRIGGER_MOVE buttons
-        def _reapply_trigger_pause(paused, button_idx):
-            is_active = state.button_active_mask[button_idx]
-            is_trigger = (
-                state.button_action_type[button_idx] == ButtonAction.TRIGGER_MOVE
-            )
-
-            def _pause_target(paused, target_slot):
-                target_idx = state.button_target_idxs[button_idx, target_slot]
-                target_enabled = state.button_target_mask[button_idx, target_slot]
-                mw_idx = jnp.clip(target_idx, 0, MAX_MOVING_WALLS - 1)
-                new_paused = jax.lax.select(
-                    is_active & is_trigger & target_enabled,
-                    paused.at[mw_idx].set(True),
-                    paused,
-                )
-                return new_paused, None
-
-            paused, _ = jax.lax.scan(
-                _pause_target, paused, jnp.arange(MAX_BUTTON_TARGETS)
-            )
-            return paused, None
-
-        new_paused, _ = jax.lax.scan(
-            _reapply_trigger_pause, new_paused, jnp.arange(MAX_BUTTONS)
-        )
-
-        # Rebuild agents with updated positions
-        new_agents = state.agents.replace(pos=Position(x=new_ag_xs, y=new_ag_ys))
-
-        return state.replace(
-            grid=new_grid,
-            agents=new_agents,
-            moving_wall_positions=new_positions,
-            moving_wall_directions=new_directions,
-            moving_wall_paused=new_paused,
-        )
-
-    # -------------------------------------------------------------------------
-    # Order Queue
-    # -------------------------------------------------------------------------
+        """Compatibility wrapper for functional moving wall updates."""
+        return move_moving_walls(state, self.config)
 
     def _process_order_queue(
         self, state: State, key: chex.PRNGKey
     ) -> Tuple[State, float]:
-        """Process order queue: generate new orders, check expirations."""
-        if not self.enable_order_queue:
-            return state, 0.0
-
-        order_types = state.order_types
-        order_expirations = state.order_expirations
-        order_active_mask = state.order_active_mask
-
-        # Decrement expirations
-        new_expirations = jnp.where(
-            order_active_mask, order_expirations - 1, order_expirations
-        )
-
-        # Check for expired orders
-        expired_mask = order_active_mask & (new_expirations <= 0)
-        num_expired = jnp.sum(expired_mask)
-        reward = num_expired * ORDER_EXPIRED_PENALTY
-
-        # Deactivate expired orders
-        new_active_mask = order_active_mask & ~expired_mask
-
-        # Maybe generate new order
-        key, subkey = jax.random.split(key)
-        should_generate = jax.random.uniform(subkey) < self.order_generation_rate
-
-        # Find first empty slot
-        empty_slots = ~new_active_mask
-        first_empty_idx = jnp.argmax(empty_slots)
-        has_empty_slot = jnp.any(empty_slots)
-
-        # Generate random order type (1 = onion soup, 2 = tomato soup if num_ingredients > 1)
-        key, subkey = jax.random.split(key)
-        new_order_type = jax.random.randint(
-            subkey, (), 1, min(self.layout.num_ingredients + 1, 3)
-        )
-
-        should_add = should_generate & has_empty_slot
-        new_order_types = jax.lax.select(
-            should_add, order_types.at[first_empty_idx].set(new_order_type), order_types
-        )
-        new_expirations = jax.lax.select(
-            should_add,
-            new_expirations.at[first_empty_idx].set(self.order_expiration_time),
-            new_expirations,
-        )
-        new_active_mask = jax.lax.select(
-            should_add, new_active_mask.at[first_empty_idx].set(True), new_active_mask
-        )
-
-        return state.replace(
-            order_types=new_order_types,
-            order_expirations=new_expirations,
-            order_active_mask=new_active_mask,
-        ), reward
-
-    # -------------------------------------------------------------------------
-    # Termination, Observations, and Spaces
-    # -------------------------------------------------------------------------
+        """Compatibility wrapper for functional order queue updates."""
+        return process_order_queue(state, key, self.config)
 
     def is_terminal(self, state: State) -> bool:
-        """Check whether state is terminal."""
-        done_steps = state.time >= self.max_steps
-        return done_steps | state.terminal
+        """Compatibility wrapper for the functional terminal check."""
+        return is_terminal(state, self.config)
 
     def get_obs(self, state: State) -> Dict[str, chex.Array]:
-        """Get observations for all agents."""
-        if not isinstance(self.observation_type, list):
-            return self.get_obs_for_type(state, self.observation_type)
-
-        all_obs = {}
-        for i, obs_type in enumerate(self.observation_type):
-            obs = self.get_obs_for_type(state, obs_type)
-            key = f"agent_{i}"
-            all_obs[key] = obs[key]
-        return all_obs
+        """Compatibility wrapper for functional observation construction."""
+        return get_obs(state, self.config)
 
     def get_obs_for_type(
         self, state: State, obs_type: ObservationType
     ) -> Dict[str, chex.Array]:
-        """Get observations for a specific observation type."""
-        if obs_type == ObservationType.DEFAULT:
-            all_obs = self.get_obs_default(state)
-        elif obs_type == ObservationType.FEATURIZED:
-            # Simplified placeholder
-            all_obs = jnp.zeros((self.num_agents,) + self.obs_shape)
-        else:
-            raise ValueError(f"Invalid observation type: {obs_type}")
-
-        def _mask_obs(obs, agent):
-            view_size = self.agent_view_size
-            pos = agent.pos
-
-            padded_obs = jnp.pad(
-                obs,
-                ((view_size, view_size), (view_size, view_size), (0, 0)),
-                mode="constant",
-                constant_values=0,
-            )
-
-            sliced_obs = jax.lax.dynamic_slice(
-                padded_obs,
-                (pos.y, pos.x, 0),
-                self.obs_shape,
-            )
-
-            return sliced_obs
-
-        if self.agent_view_size is not None:
-            all_obs = jax.vmap(_mask_obs)(all_obs, state.agents)
-
-        return {f"agent_{i}": obs for i, obs in enumerate(all_obs)}
+        """Compatibility wrapper for one functional observation encoding."""
+        return get_obs_for_type(state, obs_type, self.config)
 
     def get_obs_default(self, state: State) -> chex.Array:
-        """Get default grid-based observations."""
-        width = self.width
-        height = self.height
-        num_ingredients = self.layout.num_ingredients
-
-        static_objects = state.grid[:, :, 0]
-        ingredients = state.grid[:, :, 1]
-        static_encoding = jnp.array([
-            StaticObject.WALL,
-            StaticObject.GOAL,
-            StaticObject.POT,
-            StaticObject.RECIPE_INDICATOR,
-            StaticObject.PLATE_PILE,
-            StaticObject.ITEM_CONVEYOR,
-            StaticObject.PLAYER_CONVEYOR,
-            StaticObject.MOVING_WALL,
-            StaticObject.BUTTON,
-            StaticObject.BARRIER,
-            StaticObject.PRESSURE_PLATE,
-        ])
-        static_layers = static_objects[..., None] == static_encoding
-
-        def _ingredient_layers(ingredients):
-            shift = jnp.array([0, 1] + [2 * (i + 1) for i in range(num_ingredients)])
-            mask = jnp.array([0x1, 0x1] + [0x3] * num_ingredients)
-
-            layers = ingredients[..., None] >> shift
-            layers = layers & mask
-            return layers
-
-        recipe_indicator_mask = static_objects == StaticObject.RECIPE_INDICATOR
-        recipe_ingredients = jnp.where(recipe_indicator_mask, state.recipe, 0)
-
-        pot_timer_layer = jnp.zeros((height, width), dtype=jnp.int32)
-        for i in range(MAX_POTS):
-            y, x = state.pot_positions[i]
-            timer = state.pot_cooking_timer[i]
-            is_active = state.pot_active_mask[i]
-            pot_timer_layer = jax.lax.select(
-                is_active, pot_timer_layer.at[y, x].set(timer), pot_timer_layer
-            )
-
-        extra_layers = jnp.stack([pot_timer_layer], axis=-1)
-
-        def _agent_layers(agent):
-            pos = agent.pos
-            direction = agent.dir
-            inv = agent.inventory
-
-            pos_layers = (
-                jnp.zeros((height, width, 1), dtype=jnp.uint8)
-                .at[pos.y, pos.x, 0]
-                .set(1)
-            )
-            dir_layers = (
-                jnp.zeros((height, width, 4), dtype=jnp.uint8)
-                .at[pos.y, pos.x, direction]
-                .set(1)
-            )
-            inv_grid = jnp.zeros_like(ingredients).at[pos.y, pos.x].set(inv)
-            inv_layers = _ingredient_layers(inv_grid)
-
-            return jnp.concatenate([pos_layers, dir_layers, inv_layers], axis=-1)
-
-        def _agent_obs(agent_id):
-            agent_layers = jax.vmap(_agent_layers)(state.agents)
-            agent_layer = agent_layers[agent_id]
-            all_agent_layers = jnp.sum(agent_layers, axis=0)
-            other_agent_layers = all_agent_layers - agent_layer
-
-            ingredients_layers = _ingredient_layers(ingredients)
-            recipe_layers = _ingredient_layers(recipe_ingredients)
-
-            ingredient_pile_encoding = jnp.array(
-                [StaticObject.INGREDIENT_PILE_BASE + i for i in range(num_ingredients)]
-            )
-            ingredient_pile_layers = (
-                static_objects[..., None] == ingredient_pile_encoding
-            )
-
-            return jnp.concatenate(
-                [
-                    agent_layer,
-                    other_agent_layers,
-                    static_layers,
-                    ingredient_pile_layers,
-                    ingredients_layers,
-                    recipe_layers,
-                    extra_layers,
-                ],
-                axis=-1,
-            )
-
-        return jax.vmap(_agent_obs)(jnp.arange(self.num_agents))
+        """Compatibility wrapper for functional default observations."""
+        return get_obs_default(state, self.config)
 
     @property
     def name(self) -> str:
