+++ jaxmarl/environments/overcooked_v3/step.py
@@ -1,0 +1,119 @@
+"""Functional timestep pipeline for Overcooked V3."""
+
+from typing import Dict, Tuple
+
+import chex
+import jax
+import jax.numpy as jnp
+from jax import lax
+
+from jaxmarl.environments.overcooked_v3.agent_step import run_agent_action_phase
+from jaxmarl.environments.overcooked_v3.config import OvercookedV3Config
+from jaxmarl.environments.overcooked_v3.observations import get_obs
+from jaxmarl.environments.overcooked_v3.state import State
+from jaxmarl.environments.overcooked_v3.systems.barriers import (
+    update_barrier_timers,
+    update_pressure_plates,
+)
+from jaxmarl.environments.overcooked_v3.systems.conveyors import (
+    move_items_on_item_conveyors,
+    push_players_on_player_conveyors,
+)
+from jaxmarl.environments.overcooked_v3.systems.moving_walls import move_moving_walls
+from jaxmarl.environments.overcooked_v3.systems.orders import process_order_queue
+
+def step_overcooked_v3(
+    key: chex.PRNGKey,
+    state: State,
+    actions: Dict[str, chex.Array],
+    config: OvercookedV3Config,
+) -> Tuple[Dict[str, chex.Array], State, Dict[str, float], Dict[str, bool], Dict]:
+    """Run one environment timestep as a sequence of explicit state transforms."""
+    agent_actions = translate_action_dict_to_ordered_action_array(actions, config)
+
+    state, reward, shaped_rewards = run_agent_action_phase(
+        key, state, agent_actions, config
+    )
+    state = advance_dynamic_environment_systems(state, config)
+    state, reward = advance_order_queue_and_add_queue_reward(key, state, reward, config)
+    state = advance_time_and_update_terminal_flag(state, config)
+
+    return build_step_env_return_values(state, reward, shaped_rewards, config)
+
+def translate_action_dict_to_ordered_action_array(
+    actions: Dict[str, chex.Array], config: OvercookedV3Config
+) -> chex.Array:
+    """Convert per-agent action dictionary entries into the ordered action array."""
+    return config.action_set.take(
+        indices=jnp.array([actions[f"agent_{i}"] for i in range(config.num_agents)])
+    )
+
+def advance_dynamic_environment_systems(
+    state: State, config: OvercookedV3Config
+) -> State:
+    """Advance walls, plates, conveyors, and barrier timers after agent actions."""
+    if config.enable_moving_walls:
+        state = move_moving_walls(state, config)
+
+    if config.enable_pressure_plates:
+        state = update_pressure_plates(state, config)
+
+    if config.enable_item_conveyors:
+        state = move_items_on_item_conveyors(state, config)
+
+    if config.enable_player_conveyors:
+        state = push_players_on_player_conveyors(state, config)
+
+    return update_barrier_timers(state, config)
+
+def advance_order_queue_and_add_queue_reward(
+    key: chex.PRNGKey,
+    state: State,
+    reward: float,
+    config: OvercookedV3Config,
+) -> Tuple[State, float]:
+    """Generate and expire queued orders, adding any queue reward to the step reward."""
+    if config.enable_order_queue:
+        key, subkey = jax.random.split(key)
+        state, order_reward = process_order_queue(state, subkey, config)
+        reward = reward + order_reward
+
+    return state, reward
+
+def advance_time_and_update_terminal_flag(
+    state: State, config: OvercookedV3Config
+) -> State:
+    """Increment episode time and set the terminal flag for the new state."""
+    state = state.replace(time=state.time + 1)
+    done = is_terminal(state, config)
+    return state.replace(terminal=done)
+
+def build_step_env_return_values(
+    state: State,
+    reward: float,
+    shaped_rewards: chex.Array,
+    config: OvercookedV3Config,
+) -> Tuple[Dict[str, chex.Array], State, Dict[str, float], Dict[str, bool], Dict]:
+    """Build stopped-gradient observations, state, rewards, dones, and info."""
+    obs = get_obs(state, config)
+    done = state.terminal
+
+    rewards = {f"agent_{i}": reward for i in range(config.num_agents)}
+    shaped_rewards_dict = {
+        f"agent_{i}": shaped_reward for i, shaped_reward in enumerate(shaped_rewards)
+    }
+    dones = {f"agent_{i}": done for i in range(config.num_agents)}
+    dones["__all__"] = done
+
+    return (
+        lax.stop_gradient(obs),
+        lax.stop_gradient(state),
+        rewards,
+        dones,
+        {"shaped_reward": shaped_rewards_dict},
+    )
+
+def is_terminal(state: State, config: OvercookedV3Config) -> bool:
+    """Return whether the state has reached the episode horizon or was terminal."""
+    done_steps = state.time >= config.max_steps
+    return done_steps | state.terminal
